+++ dart/gui/all.hpp
@@ -5,6 +5,7 @@
 #include <dart/gui/export.hpp>
 #include <dart/gui/geometry.hpp>
 #include <dart/gui/interaction.hpp>
+#include <dart/gui/panel.hpp>
 #include <dart/gui/profile.hpp>
 #include <dart/gui/renderable.hpp>
 #include <dart/gui/scene.hpp>